Raise error when branch coverage information cannot be merged
The merging processes requires the same coverage information be
specified; that all line and block numbers line up and occur in the
same order. If an inconsistency is noticed (a different number of
branches per line, or block ids coming in a different order) then we
fail the merging process and raise an error.