Fix handling of BA lines in CoverageOutputGenerator
BA lines were not being correctly handled; SourceFileGenerator would
incorrectly merge entries with the same line number and the "taken"
value was frequently being interpreted as an execution count.

This fixes these issues, but merging of branch coverage information is
now more subtle. We have to check for inconsistencies in the input
files where branch data does not match up: e.g. where the number
of branches reported on a single line differ, or the block ids
or branch ids for a set of branches on a line don't match.
There isn't a lot we can do if this happens, so we just error out
quickly.

These cases should only be possible if a single coverage run covers
source files built several times by different compilers that produce
incompatible output.

Closes #11911.

PiperOrigin-RevId: 326220481