Fix branch coverage merging when line numbers differ
It may be that two SourceFileCoverage objects represent different parts
of a source file, and so contain information for different line numbers.
We should not fail to merge the branch information in these objects.

This refines the branch merging logic to operate "line by line". We
still require consistency in the inputs if both merge inputs have branch
information for the same line.