Refactor determination of predeclared environment
Previously, we determined the StarlarkBuiltinsValue and the predeclared environment in two separate places, even though the former is only used for the latter, and only in one of several cases. This CL merges the two steps into a single helper and eliminates the need for a dummy value for the other cases.

The StarlarkSemantics and StarlarkBuiltinsValue skyframe deps are still retrieved before bzl load deps are requested. I don't think it matters, but it felt better to have these two singleton values get retrieved earlier rather than later in the procesing.

RELNOTES: None
PiperOrigin-RevId: 326250045