Display remote action as running only if it is actually executing.
From #11801

> due to bazel displaying full parallelism even if actions are actually queued waiting for gRPC connections

This PR makes Bazel display running actions only if it is actually executing remotely as a first step to fix parallelism limitation.

With `--jobs=500`, the progress bar is displayed as:
```
[1 / 501] 500 actions running
```

However, the number of actual running actions by remote workers is not 500 since some actions are pending because of the limitation. With this change, the progress bar now become:
```
[1 / 501] 500 actions, 200 running
```

Closes #11922.

PiperOrigin-RevId: 326604891