Prepare for bazel to run with shrunken r8.jar
With Android Platform SDK build tools 30.0.1 the r8.jar included is shrunken (r8lib),
so some of the internal APIs are not available. Change the integration to deal with
that.

This uses reflection to access the internal APIs if available and otherwise stick to
the public API.

If a shrunked r8.jar is used, bazel will not support the configuration of "check main dex"
and "minimal main dex" flags which are not supported in the public D8 API.

This will also make it possible for bazel to pull r8.jar from Google Maven, where only
the shrunken r8.jar is published.

RELNOTES: None.
PiperOrigin-RevId: 326619633