Fix branch coverage merging when line numbers differ
https://github.com/bazelbuild/bazel/pull/11899 (5a8c461) changed
SourceFileCoverage to properly account for multiple branches on a
single line. However, it erroneously assumed that, during merges, all
input coverage objects would represent the entire source file and
required the exact same number of branches to be recorded in each
input, in the same order.

It may be that two SourceFileCoverage objects represent different parts
of a source file, and so contain information for different line numbers.
We should not fail to merge the branch information in these objects.

This refines the branch merging logic to operate "line by line". We
still require consistency in the inputs if both merge inputs have branch
information for the same line.

Closes #11932.

PiperOrigin-RevId: 326625481