Check for errors from config conditions before checking for missing values.
The resolveConfiguredTargetDependencies method currently tracks both
missing values (which will require a skyframe restart) and errors from
dependencies (which should be re-thrown). Unfortunately, it checks this
in the wrong order, which means that when a value is missing and there
is an error, the error will be swallowed and not reported.