Add a blob existence cache to RemoteExecutionCache
Bazel currently performs a full findMissingBlob call for every input of
every action that is not an immediate cache hit. This causes a
significant amount of network traffic, resulting in long build times on
low-bandwidth, high-latency connections (i.e., work from home).

In particular, for a build of TensorFlow w/ remote execution over a
slow-ish network, we see multi-minute delays on actions due to these
findMissingBlob calls, and the client is not able to saturate a
50-machine remote execution cluster.

This change carefully de-duplicates findMissingBlob calls as well as
file uploads to the remote execution cluster.

It is based on a previous simpler change that did not de-duplicate
*concurrent* calls. However, we have found that concurrent calls with
the same inputs are common - i.e., an action completes that unblocks a
large number of actions that have significant overlaps in their input
sets (e.g., protoc link completion unblocks all protoc compile
actions), and we were still seeing multi-minute delays on action
execution.

With this change, a TensorFlow build w/ remote execution is down to ~20
minutes on a 50-machine cluster, and is able to fully saturate the
cluster.

We have also seen significant improvements for remote Android app
builds.

Progress on #12113.

Change-Id: I2b86076881f99d3d4a3ad8196abf9e2b5f6e0aa7