Stop needlessly parsing WORKSPACE files from external repositories.
Fixes: https://github.com/bazelbuild/bazel/issues/11936

I have an additional test to prove the spurious error message is gone and that I still can refer to a workspace by name, in https://github.com/aiuto/repro_11936/tree/more_tests
I have not added it yet because it is not really clear what the correct behavior should be for external repos - should the WORKSPACE be valid and used, or ignored. This change brings the code in line with the documentation (ignored), but the better behavior might be different.

RELNOTES: Stop needlessly parsing WORKSPACE files from external repositories.
PiperOrigin-RevId: 327157715