Print test name to stderr on failures in bash tests.
When using `enable_errexit`, the test infrastructure will print a test name
when it is present. It does so to `stdout` rather than `stderr`, unlike the
rest of the error logging. This creates a particulary problematic scenario for
commands with errors which capture stdout:

```
variable=$(command which fails)
```

After that, `variable` will contain the test name, mixed with other output if
that was produced.

Fix the logging on errors to go to stderr. Please note that this brings the
error report string in line with what `fail` does.

PiperOrigin-RevId: 327463371