Swap to parse rcfiles before parsing the command line
The main use case we're working toward for shorthand flags is to have them defined in an rcfile and then later used on the command line. Swapping the parsing order allows for the parser to build the full list of known aliases before running into something that looks like an unknown option.

For #11750

PiperOrigin-RevId: 327485224