Use CoverageReportGenerator at HEAD in integration tests
Splits the coverage integration test into versions that test either the
released version of the coverage tool or one built at HEAD.

The java coverage test already has a "head"/"released" split for the
toolchain, which includes the java coverage generator; we include the
HEAD version of the coverage tool in the appropriate test.

We have to disable the java "head" test for the moment because the test
is not compatible with ongoing work that fixes the coverage output output.
The tests will be updated when that work is complete.

Fixes #11604

Closes #11966.

PiperOrigin-RevId: 327622112