Make PackageFactory the canonical home for Starlark env business logic
This moves the computation of the predeclared environment for bzl files from StarlarkBuiltinsFunction to PackageFactory. This will allow PackageFactory to encapsulate decisions about what symbols can and can't be overridden, reducing its API surface. StarlarkBuiltinsFunction only needs to obtain the exports dicts and call into PackageFactory. At the same time, we simplify BzlLoadFunction by moving its cache of (uninjected) predeclared environments into PackageFactory.

This change will make it easier for ASTFileLookupFunction to access the names of predeclared symbols, which will soon vary depending on the type of bzl being loaded.

(Thanks to adonovan@ for suggesting that PackageFactory be the one to manage the environment.)

PackageFactory:
- Added fields to hold the top-level environments for the three dialects of bzl files -- BUILD-loaded, WORKSPACE-loaded, and builtins-loaded. For BUILD-loaded ones, we store the uninjected environment as a field, but the injected environment must be computed dynamically. (The other dialects don't support injection.) Note that in time, BUILD-loaded and WORKSPACE-loaded bzls should converge.
- Renamed the existing two fields that hold the (uninjected) contents of "native" for consistency.
- Added create*() static initializer methods, with code and TODOs taken from StarlarkBuiltinsFunction.
- Moved the method for creating the injected environment to here from StarlarkBuiltinsFunction.

Work toward #11437

RELNOTES: None
PiperOrigin-RevId: 327659505