Fix JacocoLCOVFormatter to emit BRDA lines for branches.
We have to ignore the execution counts parameter because we do not get
that information from Jacoco. The same is true of the DA lines; we
simply record an execution count of "1" if a source line is hit.

This also updates the java coverage integration tests, enabling the
head tests and disabling the release tests. This is temporary and
they can be enabled once the updated tool is released.

Closes #11919.

PiperOrigin-RevId: 327806111