Support constraint_value directly in select().
No more need for a redundant config_setting. See #8583
for an example.

This was a bit subtle. constraint_value can't directly
export a ConfigMatchingProvider because it needs to know the
platform to determine if it's a match. But platforms are
built out of constraint_values, so the platform isn't available
yet. So the parent target with the select() provides this detail.

Also beautifies "invalid select() key" errors in support of https://github.com/bazelbuild/bazel/issues/11984.

Fixes https://github.com/bazelbuild/bazel/issues/8583.

RELNOTES[NEW]: select() directly supports constraint_value (no need for an intermediate config_setting).

PiperOrigin-RevId: 327885099
Change-Id: Ifbdfa275ff83f0cb30207a5bd77aca317599fc21