Add support for sending archived tree artifacts for tree artifacts created by shared actions.
For shared actions, Bazel will represent both actions and run only one of them.
Later, for actions which we do not run, we create a copy of the
`ActionExecutionValue` carrying the same metadata for matching artifacts. We
cannot simply reuse the value since we create duplicate artifacts in the
analysis phase which that would mismatch with the action outputs for actions we
do not run.

Add logic to include the archived representation when creating a
`TreeArtifactValue` created for a shared action. Add test for
`ActionExecutionValue` transformation for shared actions.

PiperOrigin-RevId: 327900376