Track detailed failures in ErrorSensingEventHandler, use in query
This adds functionality to ErrorSensingEventHandler to support
remembering properties of error events, alongside whether any event was
an error at all.

Note that the new functionality does not explicitly specify
DetailedExitCode, FailureDetails, or types like those, which do not
belong in the general lib/events package.

This uses the new functionality in parts of Bazel's query evaluation
infrastructure to relay details about failure modes which don't result in
the evaluation's immediate termination (via, e.g., exception; this is
common when in "keep going" mode).

Subsequent CLs extend this work to more aspects of query evaluation. This
establishes a foundation for that work.

RELNOTES: None.
PiperOrigin-RevId: 328194472