bazel packages: improve space efficiency of generator_{function,location} attributes by leveraging the recently introduced sparse representation of AttributeContainer.
Before this change, generator_* attributes were recorded explicitly, costing about 10 bytes (ignoring space to store the values) per generated rule.
When callstack is recorded, this is redundant information.
This change derives the attributes from the rule's instantiation call stack,
if and only if the --record_rule_instantiation_callstack flag is enabled.

When returning attribute value: If attribute is not explicitly stored, infer it from callstack.
During rule creation: If callstack recording is enabled, do not write generator_{location,function} attributes.

Consequences:
- This change is a functional no-op, irrespective of whether callstack recording is enabled or not.
- If callstack recording is enabled, memory consumption is reduced

NOTE: No changes to generator_name, since that is NOT available in the callstack.
PiperOrigin-RevId: 328216724