Consolidate BlazeCommandDispatcher constructors
Remove constructors that take BlazeCommands as arguments. The documentation for
them was outdated (they don't do anything with modules) and they had the
undocumented side-effect of mutating the runtime. Instead, explicitly mutate the
runtime in the places where it's required to change the commands. This isn't
pretty, but it's more clear.

PiperOrigin-RevId: 328422495