Tool paths are relative to cc_toolchain not suite
If your cc_toolchain_suite references cc_toolchains in different
packages, currently path computation is incorrect because it attempts to
append relative paths to the package containing the cc_toolchain_suite
not the cc_toolchain.

This commit makes the paths be looked up relative to the cc_toolchain.

Where these happen to be the same package (which must always have been
the case before, beacuse this code requires it), the behaviour will be
identical. But if separate packages are used, that will now work too.

Closes #11688.

PiperOrigin-RevId: 328493482