Add pipeline build scripts for titan build
Bazel tests don't necessarily work out of the box under SnapRR, or on
the Mac. While tweaking the environment to make sure they all pass is
doable, the value this provide in our context is debatable. Tests run
more than 6h to run in CI, and few (if any) errors were catched by these
tests in the past. The one test really we care about is whether or not a
particular bazel executable works with apollo and tOS.

So as we move to bazel 5, we're trimming down our pipeline to its core
by doing the following:
* only build bazel, and upload it to Artifactory
* enable remote caching
* run the linux build within SnapRR

Issue: BEN-13563  Upgrade apollo to bazel 5
Change-Id: I624529bd34b6b1be8a8d2cbc081731615163af7a
Reviewed-on: https://gerrit.sg.apple.com/c/swplatform/bazelbuild/bazel/+/345648
Tested-by: Hyperion Jenkins <hyperion_jenkins@apple.com>
Reviewed-by: Paul (SPG) Janzen <paul_janzen@apple.com>