Add a mechanism (-i) to let linux-sandbox handle SIGINTs gracefully.
Without the -i option, linux-sandbox forwards SIGINTs to the child
process.  With the -i option, SIGINTs are converted to SIGTERMs,
followed by a grace period after which the child receives a SIGKILL.

This is to match the behavior of --graceful_sigterm in the
process-wrapper and allow us to expose this "graceful interrupt" feature
to individual test rules.

PiperOrigin-RevId: 328793148