Move pid file watcher to its own class
Declutters GrpcServerImpl, makes it easier to test in isolation, and makes the
pid more readily available for other things that may want it.

The @VisibleForTesting arrangement is slightly awkward, but I think it's worth
it to have the test coverage given the number of times I messed up this
refactoring.

PiperOrigin-RevId: 328986378