bazel syntax: fix bug in comprehension resolution
Following Python3, the operand of the first for clause
is treated specially: it is resolved outside the comprehension
block.

The scope of all for variables is the entire comprehension block,
so the operand of a later for clause may refer to its own variable.
It is possible to use this dynamically for a loop-carried dependence.

Thanks to @stepancheg for identifying the problem.

PiperOrigin-RevId: 329298653