Fix multi-word test case names to use lowerCamelCase.
This is to follow the style guide recommendation in:
https://google.github.io/styleguide/javaguide.html#s5.2.3-method-names
and to fix lint warnings that have already gotten in the way of two
of my recent CLs.

This change was automatically generated by collecting the list of all
affected files in a "list" file and doing:

for f in $(cat /tmp/list); do
  sed -itmp -e '/void test.*_[A-Z]/s/\(_[A-Z]\)/\L\1/g' "$f"
done

... and then skimming through the diff to undo a few "fixes" that were
incorrect because we have test names encoding literals like "BUILD".

PiperOrigin-RevId: 329363594