Avoid downloading stdout/stderr in the UI's critical path.
The UiEventHandler was reading the contents of an action's stdout and
stderr streams within a critical section.  This is problematic because
this happens once a Skyframe thread completes action execution and
posts an event (via a Reporter) to notify the UI about this condition.
The Reporter is synchronous, which means that... if handling the event
blocks for any reason, we end up serializing all calling threads on it.
And doing I/O likely blocks.

Consider this pathological build: 10000 independent novel (uncached)
actions that print a unique message to stdout.  We run these actions
remotely, which means their output is only available on demand.  Within
Google, we fetch these outputs lazily upon first open via a FUSE file
system, which means the I/O in the UI would get stuck waiting for these
downloads.  Consequentially, all Skyframe threads would make serial
progress.

To fix this, move I/O outside of the big synchronized chunk in
UiEventHandler.

Given that this causes us to buffer the whole stdout/stderr in memory
before dumping it to the console, and because this could be problematic
if many concurrent actions try to print big outputs, add a new flag
--experimental_ui_max_stdouterr_memory_bytes to limit how big of an
output we print. This flag is set conservatively by default to 1 MiB.
Nobody will pay attention to the output of an action that exceeds this
very generous limit.

For the pathological build described above, this reduces execution time
from 360+ seconds to 27 on a machine in the office with a great Internet
connection.  On my home machine, with my more modest connection, the
execution time goes from 800+ seconds to 60.

As a side-effect of this change, we can remove all OutputReference logic
which is a nice cleanup.

RELNOTES: Bazel will now skip printing action stdout/stderr contents if
they exceed --experimental_ui_max_stdouterr_memory_bytes.
PiperOrigin-RevId: 329513065