starlark: rename binary to starlark and class to Main
This allows us to import syntax.Starlark. Weirdly Java still doesn't seem
to have a mechanism for renaming imports to avoid symbol conflicts.

Also, don't print the stack when we catch an I/O error.
The errors are still lame because java.nio for some reason
doesn't include the string "file not found" in its FileNotFoundException.

PiperOrigin-RevId: 329547864