Simplify memory management for RcFile.
RcFile uses pointers to avoid copying the source paths, which leads to some very tricky gotchas in the code. This change converts the pointers to integer indices, which are stable and easier to work with. It adds a tiny bit of overhead when using those indices, but makes the entire class much easier to work with.

Note: This change makes it theoretically possible to replace the unique_ptr in RcFile::Parse with a simpler type like std::optional, but that requires C++17 which Bazel's build doesn't currently support. There's a C++11 compatible version in absl, but that requires bringing in that extra dependency which would need to be done separately. For now, I couldn't think of a better way to handle it than leaving it a unique_ptr.

RELNOTES: none
PiperOrigin-RevId: 329574479