Fix cleanup of test streaming state on an interrupt.
There were several places in the code where we were creating objects that
wrap output streams but not closing them on an interrupt.  As a result, the
streamed output file watcher would try to write to those streams after
the Bazel command already completed (losing accesss to the console) and
crashed with:

java.lang.IllegalStateException: Stream is already completed, no further calls are allowed
        at com.google.common.base.Preconditions.checkState(Preconditions.java:508)
        at io.grpc.stub.ServerCalls$ServerCallStreamObserverImpl.onNext(ServerCalls.java:340)
        at com.google.devtools.build.lib.server.GrpcServerImpl$BlockingStreamObserver.onNext(GrpcServerImpl.java:192)
        at com.google.devtools.build.lib.server.GrpcServerImpl$RpcOutputStream.write(GrpcServerImpl.java:258)
        at com.google.devtools.build.lib.util.io.DelegatingOutErr$DelegatingOutputStream.write(DelegatingOutErr.java:117)
        at java.base/java.io.BufferedOutputStream.flushBuffer(BufferedOutputStream.java:81)
        at java.base/java.io.BufferedOutputStream.flush(BufferedOutputStream.java:142)
        at com.google.devtools.build.lib.runtime.UiEventHandler$FullyBufferedOutputStreamMaybeWithCounting.flush(UiEventHandler.java:218)
        at com.google.devtools.build.lib.runtime.UiEventHandler.handleIfCapacityPermits(UiEventHandler.java:402)
        at com.google.devtools.build.lib.runtime.UiEventHandler.handleInternal(UiEventHandler.java:502)
        at com.google.devtools.build.lib.runtime.UiEventHandler.handle(UiEventHandler.java:521)
        at com.google.devtools.build.lib.events.Reporter.handle(Reporter.java:142)
        at com.google.devtools.build.lib.events.ReporterStream.write(ReporterStream.java:55)
        at com.google.devtools.build.lib.exec.TestLogHelper$FilterTestHeaderOutputStream.write(TestLogHelper.java:123)
        at com.google.devtools.build.lib.util.io.FileWatcher.run(FileWatcher.java:96)

More interestingly, this crash was only visible with the sandboxed strategy
because its spawn runner wasn't properly waiting for interrupted subprocesses to
exit.  This gave a chance to the subprocesses to touch their stdout/stderr after
we had already shut down the connection to the Bazel client.

The way this manifested to the user is that, after pressing Ctrl+C on a running
sandboxed test, Bazel silenty crashed and a subsequent command had to bring up
the server again for no obvious reason.

PiperOrigin-RevId: 329926205