Add the 'supports-graceful-termination' tag to request nice termination.
If an action specifies this new requirement and it spawns subprocesses via
the process-wrapper or the linux-sandbox, then interrupting Bazel while the
action is running will cause these subprocesses to be terminated gracefully
(first with a SIGTERM, then with a SIGKILL).

Tests that implement cleanup routines on interrupt can benefit from this to
ensure that those cleanup steps take place.

We aren't making this the default for all actions to keep Ctrl+C fast
(which arguably is already too slow sometimes).

Note that for this to work consistently across the local and the sandboxed
strategies, I've also had to fix AbstractSandboxSpawnRunner to properly
await for interrupted subprocesses, just like we did for the
LocalSpawnRunner back in https://github.com/bazelbuild/bazel/commit/de8d4c7bddb729d3703c97c6bee58062e154ec34.

RELNOTES: Added support for a 'supports-graceful-termination' execution
requirement and tag, which causes Bazel to send a SIGTERM to any tagged
actions before sending a delayed SIGKILL. This is to give actions, and more
specifically tests, a chance to clean up after themselves.
PiperOrigin-RevId: 329940226