bazel syntax: break last Starlark dependency on Bazel
This was achieved by forking util.EvaluationTestCase and then specializing it
for the needs of lib.syntax, by deleting the parts related to lib.events
and StarlarkSemantics options parsing.

This means we have two copies of this tangled class, but I plan to make
the lib.syntax copy small enough to drown in a bathtub in follow-up changes.
The syntax.util copy will migrate into lib.packages.

PiperOrigin-RevId: 329968102