bazel packages: rename StarlarkSemanticsOptions -> packages.semantics.BuildLanguageOptions
It is not a part of lib.packages, but a dependency of it;
some clients (e.g. skydoc) depend on it alone.

FYI: syntax.StarlarkSemantics does not have a code dependency on Bazel, but
it has many logical dependencies on Bazel concepts. Follow-up changes will
make StarlarkSemantics a base class, and define
packages.semantics.BuildLanguageSemantics as a Bazel-specific subclass
containing all Bazelisms.
PiperOrigin-RevId: 330004717