Fix flaky test caused by incorrectly handled method reference
For reasons that are not entirely clear to me, using a method reference
to a variadic function as a BinaryOperator in a reduce call does not
seem to work reliably, at least on the version of Java used in RBE.
In this instance, the merge function will only receive a single argument

e.g.
[a, b, c].stream().reduce(...) results in "c" rather than "a * b * c"

Replacing the reference with a lambda that passes the arguments
explicitly avoids this. Wrapping the call in another variadic function
within the same class also works, but just looks weird and is more code
that shouldn't need to exist in the first place.

The original flaky test is made stricter so this issue could be exposed
consistently, at least on RBE.

Closes #12041.

PiperOrigin-RevId: 330373138