bazel packages: decentralize StarlarkSemantics
Prior to this change, StarlarkSemantics was a closed set of options,
which meant that it had to enumerate all the optional key/value features
of its client, Bazel. (This of course makes no sense for other clients
such as Copybara.)

This change turns it into an open and extensible set of key/value pairs.
All the Bazel semantics option have been moved to packages.semantics.BuildLanguageOptions.

This removes the last logical dependency of Starlark on Bazel.

Also:
- fix a few places where the old FlagIdentifier names were inconsistent
  with the actual flag name.
- add a static check that boolean semantics options have a name that
  starts with + or - (to specify the default value).
PiperOrigin-RevId: 330379130