Minimize stats needed to print an action's stdout/stderr.
Add caching of file sizes to FileRecordingOutputStream so that we can
minimize the number of stat system calls needed to print an action's
stdout/stderr.  The unit test that exists for this (not yet open sourced,
sorry) asserts that we still do zero stats for an empty output and that
the number drops from 5 to 1 for non-empty outputs.

I have tried to measure the performance impact of this on a pretty noisy
build running on my home Internet connection.  As expected, this change
has no quantifiable effects on overall build performance.  That said,
excessive I/O has traditionally been a problem on Macs specifically, so
I'll take any gains we can get there.

PiperOrigin-RevId: 330589882