Add a new method to `MetadataProvider` to optionally provide a `FileSystem` to use when operating on `Artifact` paths.
Some of the `ActionInput` instances are `Artifact`s which means that we have
the ability to operate on their path. In some cases, we cannot use those paths
as is since `FileSystem` redirection is needed. The new method provides that
information as well as the `FileSystem` to use in such case.

PiperOrigin-RevId: 330612369