Stop treating target patterns ending in "/BUILD" specially. InterpretPathAsTarget will do fine, and this had three or four separate bugs:
(1) If a package was named BUILD, then we couldn't refer to the //BUILD:BUILD target using the standard //BUILD shorthand syntax.
(2) For Bazel, BUILD.bazel files should be preferred, and BUILD can actually just be the name of a normal target, making (1) even worse.
(3) If there was a target that ended with "/BUILD", then it could not be referred to via the sloppy "path" syntax (//pkg/targetprefix/BUILD).
(4) BUILD files could not be referenced from subdirectories using path syntax.

We do additional work of attempting to load the package foo/BUILD now (i.e., checking to see if foo/BUILD is a directory containing a BUILD file), but it's necessary for correctness, and if someone wants to avoid that work, they just need to change their target pattern to foo:BUILD.

In the common case that the user wants the package foo/bar when specifying foo/bar/BUILD, if the file foo/bar/BUILD is missing, we'll load more segments, and then notice the error when there's no target with the right name in the first package we come to.

RELNOTES[INC]: The syntax //foo/BUILD can no longer be used on the command line to refer to the //foo:BUILD target. Use //foo:BUILD (preferred) or foo/BUILD instead. This does not affect BUILD/bzl files, where that syntax already didn't work.

Closes #7923

PiperOrigin-RevId: 331000042